+++ massiv-io/src/Data/Massiv/Array/IO/Image/JuicyPixels/TIF.hs
@@ -100,6 +100,24 @@ instance Writable TIF (Image S CM.CMYK Word8) where
 instance Writable TIF (Image S CM.CMYK Word16) where
   encodeM TIF _ img = pure $ JP.encodeTiff (toJPImageCMYK16 img)
 
+instance Writable TIF (Image S Y' Word8) where
+  encodeM f opts = encodeM f opts . demoteLumaImage
+
+instance Writable TIF (Image S Y' Word16) where
+  encodeM f opts = encodeM f opts . demoteLumaImage
+
+instance Writable TIF (Image S Y' Word32) where
+  encodeM f opts = encodeM f opts . demoteLumaImage
+
+instance Writable TIF (Image S Y' Float) where
+  encodeM f opts = encodeM f opts . demoteLumaImage
+
+instance Writable TIF (Image S (Alpha Y') Word8) where
+  encodeM f opts = encodeM f opts . demoteLumaAlphaImage
+
+instance Writable TIF (Image S (Alpha Y') Word16) where
+  encodeM f opts = encodeM f opts . demoteLumaAlphaImage
+
 instance Writable TIF (Image S (Y D65) Word8) where
   encodeM f opts = encodeM f opts . toImageBaseModel
 
@@ -181,6 +199,24 @@ instance Readable TIF (Image S CM.CMYK Word16) where
   decodeWithMetadataM = decodeWithMetadataTIF
 
 
+instance Readable TIF (Image S Y' Word8) where
+  decodeWithMetadataM f = fmap (first promoteLumaImage) . decodeWithMetadataM f
+
+instance Readable TIF (Image S Y' Word16) where
+  decodeWithMetadataM f = fmap (first promoteLumaImage) . decodeWithMetadataM f
+
+instance Readable TIF (Image S Y' Word32) where
+  decodeWithMetadataM f = fmap (first promoteLumaImage) . decodeWithMetadataM f
+
+instance Readable TIF (Image S Y' Float) where
+  decodeWithMetadataM f = fmap (first promoteLumaImage) . decodeWithMetadataM f
+
+instance Readable TIF (Image S (Alpha Y') Word8) where
+  decodeWithMetadataM f = fmap (first promoteLumaAlphaImage) . decodeWithMetadataM f
+
+instance Readable TIF (Image S (Alpha Y') Word16) where
+  decodeWithMetadataM f = fmap (first promoteLumaAlphaImage) . decodeWithMetadataM f
+
 instance Readable TIF (Image S (Y D65) Word8) where
   decodeWithMetadataM f = fmap (first fromImageBaseModel) . decodeWithMetadataM f
 
@@ -249,56 +285,37 @@ instance (Mutable r Ix2 (Pixel cs e), ColorSpace cs i e) =>
   decodeWithMetadataM = decodeAutoWithMetadataTIF
 
 encodeTIF ::
-     forall r cs e m.
-     (ColorModel cs e, Source r Ix2 (Pixel cs e), MonadThrow m)
+     forall cs e m.
+     (ColorModel cs e, MonadThrow m)
   => TIF
-  -> Image r cs e
+  -> Image S cs e
   -> m BL.ByteString
 encodeTIF f img =
-  fromMaybeEncode f (Proxy :: Proxy (Image r cs e)) $
+  fromMaybeEncode f (Proxy :: Proxy (Image S cs e)) $
   msum
-    [ do Refl <- eqT :: Maybe (cs :~: CM.Y)
+    [ do Refl <- eqT :: Maybe (e :~: Word8)
          msum
-           [ do Refl <- eqT :: Maybe (e :~: Word8)
-                pure $ JP.encodeTiff $ toJPImageY8 img
-           , do Refl <- eqT :: Maybe (e :~: Word16)
-                pure $ JP.encodeTiff $ toJPImageY16 img
-           , do Refl <- eqT :: Maybe (e :~: Word32)
-                pure $ JP.encodeTiff $ toJPImageY32 img
-           , do Refl <- eqT :: Maybe (e :~: Float)
-                pure $ JP.encodeTiff $ toJPImageYF img
+           [ JP.encodeTiff <$> maybeJPImageY8 img
+           , JP.encodeTiff <$> maybeJPImageRGB8 img
+           , do Refl <- eqT :: Maybe (cs :~: Alpha (Opaque cs))
+                msum
+                  [JP.encodeTiff <$> maybeJPImageYA8 img, JP.encodeTiff <$> maybeJPImageRGBA8 img]
+           , JP.encodeTiff <$> maybeJPImageYCbCr8 img
+           , JP.encodeTiff <$> maybeJPImageCMYK8 img
            ]
-    , do Refl <- eqT :: Maybe (cs :~: Alpha CM.Y)
+    , do Refl <- eqT :: Maybe (e :~: Word16)
          msum
-           [ do Refl <- eqT :: Maybe (e :~: Word8)
-                pure $ JP.encodeTiff $ toJPImageYA8 img
-           , do Refl <- eqT :: Maybe (e :~: Word16)
-                pure $ JP.encodeTiff $ toJPImageYA16 img
+           [ JP.encodeTiff <$> maybeJPImageY16 img
+           , JP.encodeTiff <$> maybeJPImageRGB16 img
+           , do Refl <- eqT :: Maybe (cs :~: Alpha (Opaque cs))
+                msum
+                  [JP.encodeTiff <$> maybeJPImageYA16 img, JP.encodeTiff <$> maybeJPImageRGBA16 img]
+           , JP.encodeTiff <$> maybeJPImageCMYK16 img
            ]
-    , do Refl <- eqT :: Maybe (cs :~: CM.RGB)
-         msum
-           [ do Refl <- eqT :: Maybe (e :~: Word8)
-                pure $ JP.encodeTiff $ toJPImageRGB8 img
-           , do Refl <- eqT :: Maybe (e :~: Word16)
-                pure $ JP.encodeTiff $ toJPImageRGB16 img
-           ]
-    , do Refl <- eqT :: Maybe (cs :~: Alpha CM.RGB)
-         msum
-           [ do Refl <- eqT :: Maybe (e :~: Word8)
-                pure $ JP.encodeTiff $ toJPImageRGBA8 img
-           , do Refl <- eqT :: Maybe (e :~: Word16)
-                pure $ JP.encodeTiff $ toJPImageRGBA16 img
-           ]
-    , do Refl <- eqT :: Maybe (cs :~: CM.YCbCr)
-         Refl <- eqT :: Maybe (e :~: Word8)
-         pure $ JP.encodeTiff $ toJPImageYCbCr8 img
-    , do Refl <- eqT :: Maybe (cs :~: CM.CMYK)
-         msum
-           [ do Refl <- eqT :: Maybe (e :~: Word8)
-                pure $ JP.encodeTiff $ toJPImageCMYK8 img
-           , do Refl <- eqT :: Maybe (e :~: Word16)
-                pure $ JP.encodeTiff $ toJPImageCMYK16 img
-           ]
+    , do Refl <- eqT :: Maybe (e :~: Word32)
+         JP.encodeTiff <$> maybeJPImageY32 img
+    , do Refl <- eqT :: Maybe (e :~: Float)
+         JP.encodeTiff <$> maybeJPImageYF img
     ]