+++ massiv-io/src/Data/Massiv/Array/IO/Image/Netpbm.hs
@@ -195,6 +195,10 @@ instance Readable PGM (Image S CM.Y Word16) where
   decodeWithMetadataM = decodeNetpbmImage
 
 
+instance Readable PGM (Image S Y' Word8) where
+  decodeWithMetadataM f = fmap (first promoteLumaImage) . decodeWithMetadataM f
+instance Readable PGM (Image S Y' Word16) where
+  decodeWithMetadataM f = fmap (first promoteLumaImage) . decodeWithMetadataM f
 instance Readable PGM (Image S (Y D65) Word8) where
   decodeWithMetadataM f = fmap (first fromImageBaseModel) . decodeWithMetadataM f
 instance Readable PGM (Image S (Y D65) Word16) where
@@ -207,6 +211,10 @@ instance Readable (Sequence PGM) [Image S CM.Y Word16] where
   decodeWithMetadataM = decodePPMs fromNetpbmImage
 
 
+instance Readable (Sequence PGM) [Image S Y' Word8] where
+  decodeWithMetadataM f = fmap (first (fmap promoteLumaImage)) . decodeWithMetadataM f
+instance Readable (Sequence PGM) [Image S Y' Word16] where
+  decodeWithMetadataM f = fmap (first (fmap promoteLumaImage)) . decodeWithMetadataM f
 instance Readable (Sequence PGM) [Image S (Y D65) Word8] where
   decodeWithMetadataM f = fmap (first (fmap fromImageBaseModel)) . decodeWithMetadataM f
 instance Readable (Sequence PGM) [Image S (Y D65) Word16] where