+++ massiv-io/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 * Addition of instances for `GIF` sequences in `RGB` with `Alpha` and instanced with
   `GifDisposalMethod`
+* Addition of conversion functions: `demoteLumaImage`, `promoteLumaImage`,
+  `demoteLumaAlphaImage`, `promoteLumaAlphaImage`
+* Instnces for reading and writing images in Luma: `Y'` and `Y'A`
 
 # 0.2.0